Create MarstonWestGame
The main game file for the Marston versus West fighting game, current tasks involve modifying player class to be a fully functional fighter object for the game a state machine will most likely need to be implemented to record and update fighter status i.e. fighter is in attack state, jump state, block, etc. Hitboxes (currently a giant red square when kicking) are trying to be implemented to help identify when an attack has landed on opponents