Respawn Animation/Event
Added respawning animation to be used for the book crashing down. Player control has been taken away during the event and the alpha for the character has been halved denoting invulnerability. Player invulnerability not added yet. Book sprite not yet added. The KO() function has switches in it that enable the respawnEvent() function to activate and restart it if the player falls off or dies during the invulnerability time. Variables respawnSwitch and m have been added.